# backend/internal/gen/openapi/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/google/uuid"
	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
)

// Coaster Reference to a coaster.
type Coaster struct {
	Id   UUID   `json:"id"`
	Name string `json:"name"`
}

// Credit A coaster the rider has ridden at least once, aggregated over every ride of it. The id is the coaster's id.
type Credit struct {
	FirstRiddenAt time.Time `json:"firstRiddenAt"`
	Id            UUID      `json:"id"`
	LastRiddenAt  time.Time `json:"lastRiddenAt"`
	Name          string    `json:"name"`

	// RideCount Number of logged rides on this coaster
	RideCount int `json:"rideCount"`
}

// Credits defines model for Credits.
type Credits = []Credit

// Ride A single logged trip on a coaster.
type Ride struct {
	// Coaster Reference to a coaster.
	Coaster  Coaster   `json:"coaster"`
	RiddenAt time.Time `json:"riddenAt"`
}

// RidePage defines model for RidePage.
type RidePage struct {
	// NextCursor Pass as cursor to fetch the next page. Absent on the last page.
	NextCursor *string `json:"nextCursor,omitempty"`
	Rides      Rides   `json:"rides"`
}

// Rides defines model for Rides.
type Rides = []Ride

// UUID defines model for UUID.
type UUID = uuid.UUID

// WithId defines model for WithId.
type WithId struct {
	Id UUID `json:"id"`
}

// WithName defines model for WithName.
type WithName struct {
	Name string `json:"name"`
}

// LogRideParams defines parameters for LogRide.
type LogRideParams struct {
	// Timestamp the time at which the views is being recorded
	Timestamp time.Time `form:"timestamp" json:"timestamp"`
}

// ListRidesParams defines parameters for ListRides.
type ListRidesParams struct {
	// Limit Maximum rides to return
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`

	// Cursor Opaque cursor from a previous response's nextCursor. Returns rides older than the one it points at.
	Cursor *string `form:"cursor,omitempty" json:"cursor,omitempty"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// ListCredits List distinct coasters ridden, one entry per coaster
	// (GET /credits)
	ListCredits(ctx echo.Context) error
	// LogRide log a single views on a coaster. must populate either name of coaster or coaster id.
	// (POST /log/{coasterID})
	LogRide(ctx echo.Context, coasterID UUID, params LogRideParams) error
	// ListRides List logged rides, most recent first
	// (GET /rides)
	ListRides(ctx echo.Context, params ListRidesParams) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// ListCredits converts echo context to params.
func (w *ServerInterfaceWrapper) ListCredits(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ListCredits(ctx)
	return err
}

// LogRide converts echo context to params.
func (w *ServerInterfaceWrapper) LogRide(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "coasterID" -------------
	var coasterID UUID

	err = runtime.BindStyledParameterWithOptions("simple", "coasterID", ctx.Param("coasterID"), &coasterID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter coasterID: %s", err))
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params LogRideParams
	// ------------- Required query parameter "timestamp" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "timestamp", ctx.QueryParams(), &params.Timestamp, runtime.BindQueryParameterOptions{Type: "string", Format: "date-time"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter timestamp: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.LogRide(ctx, coasterID, params)
	return err
}

// ListRides converts echo context to params.
func (w *ServerInterfaceWrapper) ListRides(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListRidesParams
	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", ctx.QueryParams(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter limit: %s", err))
	}

	// ------------- Optional query parameter "cursor" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "cursor", ctx.QueryParams(), &params.Cursor, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter cursor: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ListRides(ctx, params)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlersOptions configures RegisterHandlersWithOptions.
type RegisterHandlersOptions struct {
	// BaseURL is prepended to every registered path so the API can be served
	// under a prefix.
	BaseURL string
	// OperationMiddlewares lets the caller attach per-operation middleware at
	// registration time. The map key is the OpenAPI `operationId` value as it
	// appears in the spec (the raw, un-normalized form). Operations that have
	// no entry are registered with no extra middleware. A nil map disables
	// per-operation middleware entirely.
	OperationMiddlewares map[string][]echo.MiddlewareFunc
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{})
}

// RegisterHandlersWithBaseURL registers handlers and prepends BaseURL to the
// paths so the API can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{BaseURL: baseURL})
}

// RegisterHandlersWithOptions registers handlers using the supplied options,
// including any per-operation middleware.
func RegisterHandlersWithOptions(router EchoRouter, si ServerInterface, options RegisterHandlersOptions) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(options.BaseURL+"/credits", wrapper.ListCredits, options.OperationMiddlewares["listCredits"]...)
	router.GET(options.BaseURL+"/rides", wrapper.ListRides, options.OperationMiddlewares["listRides"]...)
	router.POST(options.BaseURL+"/log/:coasterID", wrapper.LogRide, options.OperationMiddlewares["logRide"]...)

}

type ListCreditsRequestObject struct {
}

type ListCreditsResponseObject interface {
	VisitListCreditsResponse(w http.ResponseWriter) error
}

type ListCredits200JSONResponse Credits

func (response ListCredits200JSONResponse) VisitListCreditsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type LogRideRequestObject struct {
	CoasterID UUID `json:"coasterID"`
	Params    LogRideParams
}

type LogRideResponseObject interface {
	VisitLogRideResponse(w http.ResponseWriter) error
}

type LogRide200JSONResponse Ride

func (response LogRide200JSONResponse) VisitLogRideResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type ListRidesRequestObject struct {
	Params ListRidesParams
}

type ListRidesResponseObject interface {
	VisitListRidesResponse(w http.ResponseWriter) error
}

type ListRides200JSONResponse RidePage

func (response ListRides200JSONResponse) VisitListRidesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// ListCredits List distinct coasters ridden, one entry per coaster
	// (GET /credits)
	ListCredits(ctx context.Context, request ListCreditsRequestObject) (ListCreditsResponseObject, error)
	// LogRide log a single views on a coaster. must populate either name of coaster or coaster id.
	// (POST /log/{coasterID})
	LogRide(ctx context.Context, request LogRideRequestObject) (LogRideResponseObject, error)
	// ListRides List logged rides, most recent first
	// (GET /rides)
	ListRides(ctx context.Context, request ListRidesRequestObject) (ListRidesResponseObject, error)
}

type StrictHandlerFunc func(ctx echo.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// ListCredits operation middleware
func (sh *strictHandler) ListCredits(ctx echo.Context) error {
	var request ListCreditsRequestObject

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.ListCredits(ctx.Request().Context(), request.(ListCreditsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ListCredits")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(ListCreditsResponseObject); ok {
		return validResponse.VisitListCreditsResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// LogRide operation middleware
func (sh *strictHandler) LogRide(ctx echo.Context, coasterID UUID, params LogRideParams) error {
	var request LogRideRequestObject

	request.CoasterID = coasterID
	request.Params = params

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.LogRide(ctx.Request().Context(), request.(LogRideRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "LogRide")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(LogRideResponseObject); ok {
		return validResponse.VisitLogRideResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// ListRides operation middleware
func (sh *strictHandler) ListRides(ctx echo.Context, params ListRidesParams) error {
	var request ListRidesRequestObject

	request.Params = params

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.ListRides(ctx.Request().Context(), request.(ListRidesRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ListRides")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(ListRidesResponseObject); ok {
		return validResponse.VisitListRidesResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}